refactor(input): move input validation functions to separate file
- Moved getValidCarNames() and getValidRoundCount() to InputParser.kt
- Simplified Application.kt to only focus on calling validated input
